Added: Packet OP_EqExamineInfoCmd Added: If variable equals/not equals value criteria to the struct dumper Fixed: A few packet constructors had version as uint16_t

+++ trunk/source/WorldServer/Packets/OP_CreateCharacterReplyMsg_Packet.h
@@ -28,7 +28,7 @@
 
 class OP_CreateCharacterReplyMsg_Packet : public EQ2Packet {
 public:
-	OP_CreateCharacterReplyMsg_Packet(uint16_t version)
+	OP_CreateCharacterReplyMsg_Packet(uint32_t version)
 		: EQ2Packet(version) {
 		RegisterElements();
